# src/main/java/com/example/torrent/model/api/DirectoryInputDTO.java
package com.example.torrent.model.api;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class DirectoryInputDTO {
    private String path;
    private boolean enabled;
}